add os env vars
