+++ uvoo/go-https-echo/Dockerfile
@@ -1,6 +1,6 @@
 FROM golang:bullseye
 WORKDIR /usr/src/app
-COPY https-echo.go ./
-RUN go build https-echo.go && mv https-echo /usr/local/bin/
+COPY go-https-echo.go ./
+RUN go build go-https-echo.go && mv go-https-echo /usr/local/bin/
 
-CMD ["https-echo"]
+CMD ["go-https-echo"]
